add license
Restore documentation is being added, after this license is commit is done. As this is utilizing a 3rd party, need to add license to limit liability as I'm unable to guarantee that the ACD service will maintain an open API, or even continue to allow utilization as a personal backup service. As such the BSD style disclaim is appropriate.